pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "ipykernel>=6.30.1",
    "langchain>=0.3.27",
    "langchain-core>=0.3.75",
    "langchain-google-genai>=2.1.10",
    "langgraph>=0.6.6",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "tavily-python>=0.7.11",
    "uvicorn>=0.35.0",
]
